Refactoring most of the LinkHelper code into a utility class TemplateUrl.
This feature was requested so that the URLs can be generated from the model layer as well. The reason is mostly that sometimes you want to store the URL or generate it on the fly when processing the data, for example when providing the data for a RSS feed directly from the model or when generating an URL for a tweet that is send from the model layer.